requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
cffi==1.12.2
chardet==3.0.4
Click==7.0
discord.py==1.0.0a1762+gf43690b
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==4.3.1.1.dev1
flask-nav==0.6
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
multidict==4.5.2
pycparser==2.19
PyNaCl==1.2.1
six==1.12.0
SQLAlchemy==1.3.1
visitor==0.1.3
websockets==7.0
Werkzeug==0.15.1
WTForms==2.2.1
yarl==1.3.0
